+++ controllers/research.js
@@ -18,15 +18,17 @@ const messageErrorParams = 'No ha enviado los parámetros'
 
 async function addResearch(req, res) {
 
-    var data = {name: null, description: null, category: null, id_Canton: null};
+    var data = { title: null, author: null, year: null, link: null, category: null, id_Canton: null };
     if (req.fields) {
-        data.name  = req.fields.name;
-        data.description  = req.fields.description;
-        data.category  = req.fields.category;
+        data.title = req.fields.title;
+        data.author = req.fields.author;
+        data.year = req.fields.year;
+        data.link = req.fields.link;
+        data.category = req.fields.category;
         data.id_Canton = req.fields.id_Canton;
     }
 
-    if (req.fields && req.files && ObjectId.isValid( data.id_Canton )) {
+    if (req.fields && req.files && ObjectId.isValid(data.id_Canton)) {
 
         //validate form with @hapi/joi
         const { error } = addValidation(data);
@@ -34,46 +36,49 @@ async function addResearch(req, res) {
             return responsesH.sendError(res, 400, messageErrorBody);
 
         CantonModel.findById({ _id: ObjectId(data.id_Canton) }, (err, canton) => {
-                if (err) 
-                    return responsesH.sendError(res, 400, 'Canton no encontrado.'); 
+            if (err)
+                return responsesH.sendError(res, 400, 'Canton no encontrado.');
 
-                //Upload File to the Server
-                const id_script = crypto.randomBytes(20).toString('hex');
-                var route = './images/' + id_script + '__' + req.files.image.name;
-                var tmp_path = req.files.image.path;
 
-                var readStream = fs.createReadStream(tmp_path);
-                var writeStream = fs.createWriteStream(route);
+            /*  //Upload File to the Server
+              const id_script = crypto.randomBytes(20).toString('hex');
+              var route = './images/' + id_script + '__' + req.files.image.name;
+              var tmp_path = req.files.image.path;
 
-                readStream.pipe(writeStream);                                
+              var readStream = fs.createReadStream(tmp_path);
+              var writeStream = fs.createWriteStream(route);
 
-                const research = new ResearchModel({
-                    name: req.fields.name,
-                    description: req.fields.description,
-                    category: req.fields.category,
-                    image_route: route,
-                    image_contentType: req.files.image.type,
-                    obj_Canton: canton,
-                    active: true
-                });
+              readStream.pipe(writeStream);*/
 
-                research.save((err, value) => {
-                        if (err) {
-                            try {
-                                fs.unlinkSync(route);
-                            } catch(err) {
-                        
-                            }
-                            return responsesH.sendError(res, 500, messageError);
-                        }
-                        return responsesH.sendResponseOk(res, value, 'Research insertada correctamente.');
-                });   
-                
-                try {
-                    fs.unlinkSync(tmp_path);
-                  } catch(err) {
-                    
+            const research = new ResearchModel({
+                title: req.fields.title,
+                author: req.fields.author,
+                year: req.fields.year,
+                link: req.fields.link,
+                category: req.fields.category,
+                //image_route: route,
+                //image_contentType: req.files.image.type,
+                obj_Canton: canton,
+                active: true
+            });
+
+            research.save((err, value) => {
+                if (err) {
+                    /* try {
+                         fs.unlinkSync(route);
+                     } catch (err) {
+
+                     }*/
+                    return responsesH.sendError(res, 500, messageError);
                 }
+                return responsesH.sendResponseOk(res, value, 'Research insertada correctamente.');
+            });
+
+            /*try {
+                fs.unlinkSync(tmp_path);
+            } catch (err) {
+
+            }*/
         });
     } else {
         return responsesH.sendError(res, 500, messageErrorBody);
@@ -82,15 +87,19 @@ async function addResearch(req, res) {
 
 async function updateResearch(req, res) {
 
-    var data = {name: null, description: null, category: null, id_Canton: null};
+    var data = { title: null, author: null, year: null, link: null, category: null, id_Canton: null };
     if (req.fields) {
-        data.name  = req.fields.name;
-        data.description  = req.fields.description;
-        data.category  = req.fields.category;
+        data.title = req.fields.title;
+        data.author = req.fields.author;
+        data.year = req.fields.year;
+        data.link = req.fields.link;
+        data.category = req.fields.category;
         data.id_Canton = req.fields.id_Canton;
+        //data.active = req.fields.active
+
     }
 
-    if (req.fields && req.files && ObjectId.isValid( data.id_Canton )) {
+    if (req.fields && req.files && ObjectId.isValid(data.id_Canton)) {
 
         //validate form with @hapi/joi
         const { error } = updateValidation(data);
@@ -98,23 +107,29 @@ async function updateResearch(req, res) {
             return responsesH.sendError(res, 400, messageErrorBody);
 
         CantonModel.findById({ _id: ObjectId(data.id_Canton) }, (err, canton) => {
-                if (err) 
-                    return responsesH.sendError(res, 400, 'Canton no encontrado.'); 
+            if (err)
+                return responsesH.sendError(res, 400, 'Canton no encontrado.');
 
-                const _id = req.params.id;
-                //Search Research to update and update
-                ResearchModel.findById(_id, (err, research) => {
-                    if (err) {
-                        return responsesH.sendError(res, 500, 'Research no encontrada.');
-                    }
+            const _id = req.params.id;
+            //Search Research to update and update
+            ResearchModel.findById(_id, (err, research) => {
+                if (err) {
+                    return responsesH.sendError(res, 500, 'Research no encontrada.');
+                }
 
-                    var route_to_delete = research.image_route;
+                var route_to_delete = research.image_route;
 
-                    //Update Clasification
-                    if (req.fields.name)         research.name = req.fields.name;
-                    if (req.fields.description)  research.description = req.fields.description;
-                    if (req.fields.category)     research.category = req.fields.category;
-                    if (req.fields.id_Canton)    research.canton = canton;
+                //Update Clasification
+                if (req.fields.title) research.title = req.fields.title;
+                if (req.fields.author) research.author = req.fields.author;
+                if (req.fields.year) research.year = req.fields.year;
+                if (req.fields.link) research.link = req.fields.link;
+                if (req.fields.category) research.category = req.fields.category;
+                if (req.fields.id_Canton) research.canton = canton;
+                if (req.fields.active) research.active = req.fields.active;
+
+
+                if (req.files.image) {
 
 
                     //Upload File to the Server
@@ -129,24 +144,30 @@ async function updateResearch(req, res) {
 
                     research.image_route = route;
                     research.image_contentType = req.files.image.type;
-
-                    research.save((err, value) => {
-                            if (err) {
-                                return responsesH.sendError(res, 500, 'Error actualizando la Investigación.');
-                            }
-                            return responsesH.sendResponseOk(res, value, 'Investigación actualizada correctamente.');
-                    });    
-
-                    // Delete temporary file
-                    try {
-                        fs.unlinkSync(tmp_path);
-                        fs.unlinkSync(route_to_delete);
-                      } catch(err) {
-                        
+                }
+                research.save((err, value) => {
+                    if (err) {
+                        return responsesH.sendError(res, 500, 'Error actualizando la Investigación.');
                     }
-
+                    return responsesH.sendResponseOk(res, value, 'Investigación actualizada correctamente.');
                 });
-                
+
+                // Delete temporary file
+                try {
+                    if (fs.existsSync(tmp_path)) {
+                        fs.unlinkSync(tmp_path);
+
+                    }
+                    if (fs.existsSync(route_to_delete)) {
+                        fs.unlinkSync(route_to_delete);
+
+                    }
+                } catch (err) {
+
+                }
+
+            });
+
         });
     } else {
         return responsesH.sendError(res, 500, messageErrorBody);
@@ -155,10 +176,10 @@ async function updateResearch(req, res) {
 
 async function deleteResearch(req, res) {
     _id = req.params.id
-    if ( ObjectId.isValid( _id ) ) {
+    if (ObjectId.isValid(_id)) {
 
-        const research = await ResearchModel.findOne({_id: _id});
-        if (research == null) 
+        const research = await ResearchModel.findOne({ _id: _id });
+        if (research == null)
             return responsesH.sendError(res, 500, 'Research no encontrada.');
 
         ResearchModel.deleteOne({ _id: _id }, (err, value) => {
@@ -171,8 +192,8 @@ async function deleteResearch(req, res) {
         //Delete the temporary file
         try {
             fs.unlinkSync(research.image_route);
-          } catch(err) {
-            
+        } catch (err) {
+
         }
 
     } else {
@@ -180,15 +201,17 @@ async function deleteResearch(req, res) {
     }
 }
 
-function getResearchsByCatAndCant(req,res) {
+function getResearchsByCatAndCant(req, res) {
 
     const id_Canton = req.body.id_Canton;
     const category = req.body.category;
+    const sort = { year: -1 };
 
-    if ( ObjectId.isValid(id_Canton) ) {
-        var search = {'obj_Canton._id': ObjectId(id_Canton)};
-    
-        if ( category ) {
+
+    if (ObjectId.isValid(id_Canton)) {
+        var search = { 'obj_Canton._id': ObjectId(id_Canton) };
+
+        if (category) {
             search.category = category;
         }
 
@@ -197,7 +220,7 @@ function getResearchsByCatAndCant(req,res) {
                 return responsesH.sendError(res, 500, messageError);
             }
             return responsesH.sendResponseOk(res, value);
-        });
+        }).sort(sort);
     } else {
         return responsesH.sendError(res, 500, messageErrorBody);
     }
